go from org-babel to individual files
The org babel approach to init files is a really nice idea and I'm quite
sure I will switch to that eventually again. But currently having the
init system split up into several files is easier for me to navigate and
see all the settings that I have made. Plus doing this now is a good
chance to clean it up.